+++ ydb/src/grpc_wrapper/raw_table_service/mod.rs
@@ -5,7 +5,6 @@ pub(crate) mod execute_scheme_query;
 pub(crate) mod keepalive;
 pub(crate) mod rollback_transaction;
 pub(crate) mod value;
-pub(crate) mod value_type;
 pub(crate) mod execute_data_query;
 pub(crate) mod transaction_control;
 pub(crate) mod query_stats;
