+++ ydb/src/grpc_wrapper/raw_table_service/value/value_proto.rs
@@ -1,54 +1,15 @@
-use crate::grpc_wrapper::raw_errors::RawError;
-use crate::grpc_wrapper::raw_table_service::value_type::{decode_err, RawType};
-use ydb_grpc::ydb_proto::value::Value as Primitive;
 use ydb_grpc::ydb_proto::Value as ProtoValue;
-
-#[cfg(test)]
-#[path = "value_test.rs"]
-mod value_test;
-
-#[derive(Clone, Debug, PartialEq)]
-pub(crate) struct RawTypedValue {
-    pub r#type: RawType,
-    pub value: RawValue,
-}
-
-#[derive(Clone, Debug, PartialEq, strum::EnumCount)]
-pub(crate) enum RawValue {
-    Bool(bool),
-    Int32(i32),
-    UInt32(u32),
-    Int64(i64),
-    UInt64(u64),
-    HighLow128(u64, u64), // high, low
-    Float(f32),
-    Double(f64),
-    Bytes(Vec<u8>),
-    Text(String),
-    NullFlag,
-    // NestedValue(Box<Value>), return as Variant with 0 index
-    Items(Vec<RawValue>),
-    Pairs(Vec<RawValuePair>),
-    Variant(Box<RawVariantValue>),
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub(crate) struct RawValuePair {
-    key: RawValue,
-    payload: RawValue,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub(crate) struct RawVariantValue {
-    value: RawValue,
-    index: u32,
-}
+use ydb_grpc::ydb_proto::value::Value as Primitive;
+use crate::grpc_wrapper::raw_errors::RawError;
+use crate::grpc_wrapper::raw_table_service::value::{RawColumn, RawResultSet, RawTypedValue, RawValue, RawValuePair};
+use crate::grpc_wrapper::raw_table_service::value::value_type::{decode_err, RawType};
 
 impl TryFrom<ProtoValue> for RawValue {
     type Error = RawError;
 
     fn try_from(value: ProtoValue) -> Result<Self, Self::Error> {
-        use RawValue::*;
+        use crate::grpc_wrapper::raw_table_service::value::RawValue::*;
+        use crate::grpc_wrapper::raw_table_service::value::RawVariantValue;
 
         if let Some(simple) = value.value {
             let res = match simple {
@@ -117,10 +78,6 @@ impl TryFrom<ydb_grpc::ydb_proto::ValuePair> for RawValuePair {
         })
     }
 }
-
-//
-// internal to protobuf
-//
 
 impl From<RawValue> for ProtoValue {
     fn from(v: RawValue) -> Self {
@@ -207,11 +164,6 @@ impl From<RawTypedValue> for ydb_grpc::ydb_proto::TypedValue {
     }
 }
 
-pub(crate) struct RawResultSet {
-    pub columns: Vec<RawColumn>,
-    pub rows: Vec<Vec<RawValue>>,
-}
-
 impl TryFrom<ydb_grpc::ydb_proto::ResultSet> for RawResultSet {
     type Error = RawError;
 
@@ -235,11 +187,6 @@ impl TryFrom<ydb_grpc::ydb_proto::ResultSet> for RawResultSet {
 
         Ok(Self { columns, rows })
     }
-}
-
-pub(crate) struct RawColumn {
-    pub name: String,
-    pub column_type: RawType,
 }
 
 impl TryFrom<ydb_grpc::ydb_proto::Column> for RawColumn {
